dj/requirements.txt
Django
django-allauth
django-browser-reload
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-environ
django-extensions
django-filter
django-formtools
django-spaghetti-and-meatballs
django-storages
django-taggit
djangorestframework
drf-spectacular
gunicorn
ipython
isort
itemadapter
itemloaders
jupyter
notebook
numpy
opencv-python
Pillow
psycopg2-binary
pydantic
requests-html
Scrapy
scrapy-djangoitem
uvicorn
whitenoise